+++ R/recipes-step_smooth.R
@@ -84,13 +84,11 @@
 #'
 #' @seealso
 #'  Time Series Analysis:
-#'  - [step_timeseries_signature()]
-#'  - [step_holiday_signature()]
-#'  - [step_diff()]
-#'  - [recipes::step_lag()]
-#'  - [step_roll_apply()]
-#'  - [step_smooth()]
-#'  - [step_box_cox()]
+#'  - Engineered Features: [step_timeseries_signature()], [step_holiday_signature()], [step_fourier()]
+#'  - Diffs & Lags [step_diff()], [recipes::step_lag()]
+#'  - Smoothing: [step_roll_apply()], [step_smooth()]
+#'  - Variance Reduction: [step_box_cox()]
+#'  - Imputation: [step_impute_ts()]
 #'
 #'  Main Recipe Functions:
 #'  - [recipes::recipe()]